backend/requirements.txt
# Core dependencies
Flask==2.3.3
Werkzeug==2.3.7
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0

# API and model dependencies
groq==0.4.1
requests==2.31.0
numpy==1.25.2
Pillow==10.0.0
tensorflow==2.15.0
flask-restx==1.1.0

# Development dependencies
pytest==7.4.0
pytest-flask==1.2.0
coverage==7.3.2